Revert "Merge branch 'feat-efficient-image-resolution-selection'"
The following are the reasons as to why this feature was reverted.
1) In some cases, the Palette api extracts completely different colors
for different resolutions of the same image.  This is understandable
as lower resolution images have fewer data points.

2) Since most screens require high quality images anyways, the performance
benifit that is obtained by using small images is negligible.

3) All images are cached by coil already, so the user's internet
consumption does not get affected anyways. Since, larger version
of the images have to be downloaded and saved anyways, using smaller
images don't result in improving the network consumption of the app
in a significant manner.

4) Many high-end devices, have high-resolution screens, which makes
it impractical to use low-res images since it degrades UX.

5) Since the performance and network consumption benefits are negligible,
this change only results in a lot of boilerplate code. Hence, this
merge is reverted.
This reverts commit f3eb9bf079bd638e7233138c8e63b328fcd7ec06, reversing
changes made to dfc7544254649e20f9ecca01bfc895195fa19d84.